move chart tooltip into its own folder w/ named export
