+++ src/components/data-viz/ChartTooltip/index.js
@@ -1,0 +1,1 @@
+export { default as ChartTooltip } from './ChartTooltip'
